+++ frontend/src/api/utils.test.js
@@ -54,10 +54,10 @@ describe('adjustedData', () => {
       folders: [],
       files: [],
       items: [
-        { name: "folder1", path: "/root/folder1/", source: undefined, type: "directory" },
-        { name: "folder2", path: "/root/folder2/", source: undefined, type: "directory" },
-        { name: "file1.txt", path: "/root/file1.txt", source: undefined, type: "file" },
-        { name: "file2.txt", path: "/root/file2.txt", source: undefined, type: "file" },
+        { isShared: false, name: "folder1", path: "/root/folder1/", source: undefined, type: "directory" },
+        { isShared: false, name: "folder2", path: "/root/folder2/", source: undefined, type: "directory" },
+        { isShared: false, name: "file1.txt", path: "/root/file1.txt", source: undefined, type: "file" },
+        { isShared: false, name: "file2.txt", path: "/root/file2.txt", source: undefined, type: "file" },
       ],
       path: "/root/",
     };